Split message handlers into separate files
