+++ src/components/web/WebApp.tsx
@@ -23,7 +23,8 @@ import StaticServer from 'react-native-static-server'
 import RNFS from 'react-native-fs'
 
 import PullToRefresh from './PullToRefresh'
-import { Message, handleMessage, MessageType } from '../../message'
+import { handleMessage } from '../../message/handleMessage'
+import { Message, MessageType } from '../../message/types'
 import {
   WebViewMessage,
   WebViewNavigation
@@ -565,10 +566,9 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   onMessage: (
     message: Message,
     onPostMessage: (message: Message) => void,
-    reload: () => void,
-    state: AppState
+    reload: () => void
   ) => {
-    handleMessage(message, dispatch, onPostMessage, reload, state)
+    handleMessage(message, dispatch, onPostMessage, reload)
   }
 })
