+++ src/types/analytics.ts
@@ -1,5 +1,5 @@
 import { JsonMap } from '@segment/analytics-react-native'
-import { Message } from '../message'
+import { Message } from '../message/types'
 
 export type Identify = {
   handle: string
